+++ src/main/java/mod/emt/thaumictweaker/proxy/CommonProxy.java
@@ -1,6 +1,8 @@
 package mod.emt.thaumictweaker.proxy;
 
 import mod.emt.thaumictweaker.compat.crafttweaker.CrafttweakerTT;
+import mod.emt.thaumictweaker.config.ConfigTweaksTT;
+import mod.emt.thaumictweaker.events.RunicShieldingHandler;
 import mod.emt.thaumictweaker.util.libs.ModIds;
 import net.minecraftforge.common.MinecraftForge;
 
@@ -8,6 +10,9 @@ public class CommonProxy {
     public void preInit() {
         if(ModIds.crafttweaker.isLoaded) {
             MinecraftForge.EVENT_BUS.register(new CrafttweakerTT());
+        }
+        if(ConfigTweaksTT.runic_shielding.newRunicShielding) {
+            MinecraftForge.EVENT_BUS.register(new RunicShieldingHandler());
         }
     }